poalris: overlay: reduce brightness change tolerance
The default value of 0.001f is for a panel backlight brightness
range of 0 to 255, which is approximately 1/(255*3). The LCD panels
on polaris have a max backlight brightness of 4095, so the screen
brightness cannot be adjusted fine enough with 0.001f tolerance.

Set it to 0.0001f instead.

More explanations on the framework side commit:
https://gerrit.pixelexperience.org/c/frameworks_base/+/13507

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>