polaris: ship displayconfig for brightness ramp rate
Android 12 allows the brightness ramp rate to be changed separately
for fast/slow and increase/decrease in float. Once loaded,
DisplayDeviceConfig has higher priority than the integer config
values in frameworks. The format definition is in [1].

There are 3 variants for the filename:

* display_id_46xxxxx.xml
* display_19xxxxx.xml
* display_port_xxx.xml

The value can be found in boot logs by grepping DisplayDeviceInfo.
On android 11, boot log reports the 19xxxxx value of the second
variant. On android 12, boot log reports the 46xxxxx value of the
first variant.

The xml file can be put in either product or vendor.

JDI panel: 19260591652815745 / 4630946610080203649
EBBG panel: 19261151497321857

[1] frameworks/base/services/core/xsd/display-device-config/display-device-config.xsd

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>